pdf sum report
A pdf report is now generated based on all users in the database. Follows same structure as group reports, but has locations tables placed at the bottom of the pdf since the list is so large. Further work required to change locations to a county based system rather than individual cities/states.