error update
Instead of removing the 'report' and 'group_report' folders when the code is ran, pdf files in both folders are simply deleted to prevent unnecessary errors and code breaks. Error will only be thrown now if a pdf report is open and must be closed. User can still browse the folder itself when the code is ran, but reports must be closed for new reports to be generated. Basically, code is more efficient and user friendly :)