cleaning update
Updated how the "read_users" function collects and stores user data in class objects. Instead of a direct, brute force way of collecting and storing the data, process is cleaner and is mainly done through simple for loop and dictionary. This cleaned the overall code of the function quite a bit and allows for easier alteration in the future if need be.